+++ piou/__init__.py
@@ -1,4 +1,4 @@
-from .cli import Cli, CommandGroup # noqa
-from .utils import Option, Derived, Password # noqa
-from .command import CommandMeta # noqa
-from .exceptions import CommandNotFoundError # noqa
+from .cli import Cli, CommandGroup  # noqa
+from .utils import Option, Derived, Password  # noqa
+from .command import CommandMeta  # noqa
+from .exceptions import CommandNotFoundError  # noqa