+++ tests/test_formatter.py
@@ -1,4 +1,7 @@
 import pytest
+from typing import Literal, get_origin
+
+from piou.utils import Password, CommandOption, convert_to_type
 
 
 @pytest.mark.parametrize('options, input, expected', [
@@ -14,6 +17,21 @@ def test_rich_formatting_options(options, input, expected, capsys):
     formatter = RichFormatter(**options)
     formatter._print_description(Command('', lambda x: ..., description=input))
     assert capsys.readouterr().out.rstrip() == expected
+
+
+@pytest.mark.parametrize('default, data_type,show_default,expected', [
+    pytest.param('hello', Password, True, '(default: ******)', id='Password'),
+    pytest.param(None, Literal['foo', 'bar'], True, '(choices are: foo, bar)', id='Small Literal'),
+    pytest.param(None, Literal['foo', 'bar', 'baz', 'others'], True,
+                 "\nPossible choices are: \n - foo \n - bar \n - baz \n - others", id='Large Literal')
+])
+def test_fmt_help(default, data_type, show_default, expected):
+    from piou.formatter.rich_formatter import fmt_help
+    from piou.utils import CommandOption
+    option = CommandOption(default)
+    option.data_type = data_type
+    output = fmt_help(option, show_default, markdown_open=None, markdown_close=None)
+    assert output == expected
 
 
 def get_simple_cli(formatter):
@@ -274,21 +292,28 @@ DESCRIPTION
 """
 
 _PARAMS = [
-    ('Simple CLI', get_simple_cli, ['-h'], _SIMPLE_CLI_OUTPUT_RICH),
-    ('Simple CLI cmd', get_simple_cli, ['foo', '-h'], _SIMPLE_CLI_COMMAND_RICH),
-    ('Simple CLI with opts', get_simple_cli_with_global_opt, ['-h'], _SIMPLE_CLI_WITH_OPTS_OUTPUT_RICH),
-    ('Simple CLI with opts cmd', get_simple_cli_with_global_opt, ['foo', '-h'],
-     _SIMPLE_CLI_WITH_OPTS_CMD_OUTPUT_RICH),
-    ('Simple CLI sub-cmd', get_cmd_group_cli_with_global_opt, ['-h'], _SIMPLE_CLI_SUB_CMD_RICH),
-    ('Simple CLI sub-cmd help', get_cmd_group_cli_with_global_opt, ['sub-cmd', '-h'], _SIMPLE_CLI_SUB_CMD_HELP_RICH),
-    ('Simple CLI sub-cmd 2 help', get_cmd_group_cli_with_global_opt, ['sub-cmd2', '-h'], _SIMPLE_CLI_SUB_CMD2_HELP),
-    ('Simple CLI sub-cmd cmd', get_cmd_group_cli_with_global_opt, ['sub-cmd', 'foo', '-h'],
-     _SIMPLE_CLI_SUB_CMD_CMD_RICH),
+    pytest.param(get_simple_cli, ['-h'], _SIMPLE_CLI_OUTPUT_RICH, id='Simple CLI'),
+    pytest.param(get_simple_cli, ['foo', '-h'], _SIMPLE_CLI_COMMAND_RICH, id='Simple CLI cmd'),
+    pytest.param(get_simple_cli_with_global_opt, ['-h'], _SIMPLE_CLI_WITH_OPTS_OUTPUT_RICH,
+                 id='Simple CLI with opts'),
+    pytest.param(get_simple_cli_with_global_opt, ['foo', '-h'],
+                 _SIMPLE_CLI_WITH_OPTS_CMD_OUTPUT_RICH, id='Simple CLI with opts cmd', ),
+    pytest.param(get_cmd_group_cli_with_global_opt, ['-h'], _SIMPLE_CLI_SUB_CMD_RICH,
+                 id='Simple CLI sub-cmd'),
+    pytest.param(get_cmd_group_cli_with_global_opt, ['sub-cmd', '-h'], _SIMPLE_CLI_SUB_CMD_HELP_RICH,
+                 id='Simple CLI sub-cmd help'),
+    pytest.param(get_cmd_group_cli_with_global_opt, ['sub-cmd2', '-h'], _SIMPLE_CLI_SUB_CMD2_HELP,
+                 id='Simple CLI sub-cmd 2 help'),
+    pytest.param(get_cmd_group_cli_with_global_opt, ['sub-cmd', 'foo', '-h'],
+                 _SIMPLE_CLI_SUB_CMD_CMD_RICH,
+                 id='Simple CLI sub-cmd cmd'),
     # # Errors
-    ('Simple CLI keyword error', get_simple_cli, ['foo', '-vvv'],
-     "Could not find keyword parameter '-vvv' for command 'foo'"),
-    ('Simple CLI keyword error', get_simple_cli, ['foo'],
-     "Expected 1 positional arguments but got 0 for command foo"),
+    pytest.param(get_simple_cli, ['foo', '-vvv'],
+                 "Could not find keyword parameter '-vvv' for command 'foo'",
+                 id='Simple CLI keyword error'),
+    pytest.param(get_simple_cli, ['foo'],
+                 "Expected 1 positional arguments but got 0 for command foo",
+                 id='Simple CLI keyword error'),
 ]
 
 
@@ -300,8 +325,8 @@ def _compare_str(output, expected):
         assert output_line.rstrip() == expected_line.rstrip(), print(output)
 
 
-@pytest.mark.parametrize('name, cli_fn, args, expected', _PARAMS, ids=[x[0] for x in _PARAMS])
-def test_rich_formatting(name, cli_fn, args, expected, capsys):
+@pytest.mark.parametrize('cli_fn, args, expected', _PARAMS)
+def test_rich_formatting(cli_fn, args, expected, capsys):
     from piou.formatter import RichFormatter
     cli = cli_fn(RichFormatter(show_default=False))
     cli.run_with_args(*args)